make shared versions of Avatar and AvatarCard
