+++ webui/react/src/components/NavigationSideBar.tsx
@@ -4,6 +4,7 @@ import React, { useCallback, useEffect, useMemo, useRef,
 import { useLocation } from 'react-router-dom';
 import { CSSTransition } from 'react-transition-group';
 
+import AvatarCard from 'components/AvatarCard';
 import Dropdown, { Placement } from 'components/Dropdown';
 import DynamicIcon from 'components/DynamicIcon';
 import Link, { Props as LinkProps } from 'components/Link';
@@ -15,7 +16,6 @@ import useSettings, { BaseType, SettingsConfig } from 'hooks/useSettings';
 import { clusterStatusText } from 'pages/Clusters/ClustersOverview';
 import WorkspaceActionDropdown from 'pages/WorkspaceList/WorkspaceActionDropdown';
 import { paths } from 'routes/utils';
-import AvatarCard from 'shared/components/AvatarCard/AvatarCard';
 
 import Icon from '../shared/components/Icon/Icon';
