+++ webui/react/src/pages/WorkspaceList/WorkspaceCard.tsx
@@ -2,10 +2,10 @@ import { PushpinOutlined } from '@ant-design/icons';
 import { Tooltip, Typography } from 'antd';
 import React, { useCallback } from 'react';
 
+import Avatar from 'components/Avatar';
 import DynamicIcon from 'components/DynamicIcon';
 import Link from 'components/Link';
 import { paths } from 'routes/utils';
-import Avatar from 'shared/components/Avatar/Avatar';
 import Icon from 'shared/components/Icon/Icon';
 import { routeToReactUrl } from 'shared/utils/routes';
 import { DetailedUser, Workspace } from 'types';
