+++ webui/react/src/shared/components/Avatar/Avatar.tsx
@@ -1,22 +1,15 @@
 import { Tooltip } from 'antd';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 
-import { useStore } from 'contexts/Store';
-import { useFetchUsers } from 'hooks/useFetch';
 import { hex2hsl, hsl2str } from 'shared/utils/color';
 import md5 from 'shared/utils/md5';
-import { getDisplayName } from 'utils/user';
 
 import css from './Avatar.module.scss';
 
-interface Props {
+export interface Props {
+  displayName: string;
   hideTooltip?: boolean;
   large?: boolean;
-  name?: string;
-  // TODO: separate components for
-  // 1) displaying an abbreviated string as an Avatar and
-  // 2) finding user by userId in the store and displaying string Avatar or profile image
-  userId?: number;
 }
 
 const getInitials = (name = ''): string => {
@@ -39,23 +32,7 @@ const getColor = (name = ''): string => {
   return hsl2str({ ...hslColor, l: 50 });
 };
 
-const Avatar: React.FC<Props> = ({ hideTooltip, name, large, userId }: Props) => {
-  const [ displayName, setDisplayName ] = useState('');
-  const { users } = useStore();
-  const fetchUsers = useFetchUsers(new AbortController());
-
-  useEffect(() => {
-    if (!name && userId) {
-      if (!users.length) {
-        fetchUsers();
-      }
-      const user = users.find(user => user.id === userId);
-      setDisplayName(getDisplayName(user));
-    } else if (name) {
-      setDisplayName(name);
-    }
-  }, [ fetchUsers, userId, name, users ]);
-
+const Avatar: React.FC<Props> = ({ hideTooltip, large, displayName }) => {
   const style = { backgroundColor: getColor(displayName) };
   const classes = [ css.base ];
   if (large) classes.push(css.large);
